[MOPS-4] Update deletions for existing entities (#5)

+++ src/main/java/org/example/hello/service/MealService.java
@@ -61,7 +61,15 @@ public class MealService {
         return repository.save(Meal);
     }
 
-    public void deleteMealById(String mealId) {
+    public Meal deleteMealById(String mealId) {
+        Optional<Meal> optionalMeal = repository.findById(mealId);
+        if (optionalMeal.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Meal not found.");
+        }
+
+        Meal meal = optionalMeal.get();
+
         repository.deleteById(mealId);
+        return meal;
     }
 }
